Added charts and formatting fixes
The Javascript for Chart.js and it's functionality have been added to the page.  I'm having some issues with the donut chart - there's an "uncaught TypeError" and a source map error.  I'm not sure what's causing them.
Several formatting changes have been made to the CSS to improve the layout of the page.